and move tests over (uh, completes #159)
