+++ core/src/main/java/ai/doc/tensorio/core/data/Dequantizer.java
@@ -18,7 +18,7 @@
  * limitations under the License.
  */
 
-// TODO: Dequantiziation should take bytes not ints but there are is no unsigned byte primitive (#28)
+// TODO: Dequantization should take bytes not ints but there are is no unsigned byte primitive (#28)
 
 package ai.doc.tensorio.core.data;
