+++ core/src/main/java/ai/doc/tensorio/core/modelbundle/JSONParsing.java
@@ -24,8 +24,6 @@ import ai.doc.tensorio.core.layerinterface.DataType;
 import ai.doc.tensorio.core.layerinterface.StringLayerDescription;
 import ai.doc.tensorio.core.model.ImageVolume;
 import ai.doc.tensorio.core.model.PixelFormat;
-import ai.doc.tensorio.core.modelbundle.ModelBundle;
-import ai.doc.tensorio.core.modelbundle.ModelBundleException;
 import ai.doc.tensorio.core.utilities.AndroidAssets;
 import ai.doc.tensorio.core.utilities.FileIO;
 import androidx.annotation.NonNull;
