+++ core/src/main/java/ai/doc/tensorio/core/modelbundle/Manager.java
@@ -33,8 +33,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
-import ai.doc.tensorio.core.modelbundle.ModelBundle;
-import ai.doc.tensorio.core.modelbundle.ModelBundleException;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
