+++ tflite/src/androidTest/assets/1_in_1_out_pixelbuffer_test.tiobundle/model.json
@@ -1,0 +1,43 @@
+{
+	"name": "1 in 1 out pixel buffer test",
+	"details": "Takes two pixel buffers, treating the first as a new input and the second as a previous input, and applies an exponential decay.",
+	"id": "1_in_1_out_pixelbuffer_test",
+	"version": "1",
+	"author": "doc.ai",
+	"license": "",
+	"model": {
+		"file": "model.tflite",
+		"quantized": false
+	},
+	"inputs": [
+		{
+			"name": "new_image",
+			"type": "image",
+			"shape": [224,224,3],
+			"format": "RGB",
+			"normalize": {
+				"standard": "[0,1]"
+			}
+		},
+		{
+			"name": "previous_image",
+			"type": "image",
+			"shape": [224,224,3],
+			"format": "RGB",
+			"normalize": {
+				"standard": "[0,1]"
+			}
+		}
+	],
+	"outputs": [
+		{
+			"name": "previous_image",
+			"type": "image",
+			"shape": [224,224,3],
+			"format": "RGB",
+			"denormalize": {
+				"standard": "[0,1]"
+			}
+		}
+	]
+}
