+++ tflite/src/androidTest/assets/1_in_1_out_vectors_test.tiobundle/model.json
@@ -1,0 +1,26 @@
+{
+	"name": "1 in 1 out vectors test",
+	"details": "Simple model with a single vector valued input and single vector valued output",
+	"id": "1_in_1_out_vectors_test",
+	"version": "1",
+	"author": "doc.ai",
+	"license": "",
+	"model": {
+		"file": "model.tflite",
+		"quantized": false
+	},
+	"inputs": [
+		{
+			"name": "input",
+			"type": "array",
+			"shape": [4]
+		}
+	],
+	"outputs": [
+		{
+			"name": "output",
+			"type": "array",
+			"shape": [4]
+		}
+	]
+}
