+++ tflite/src/androidTest/assets/2_in_2_out_matrices_test.tiobundle/model.json
@@ -1,0 +1,36 @@
+{
+	"name": "1 in 2 out matrices test",
+	"details": "Simple model with two matrix input nodes and two matrix output nodes",
+	"id": "2_in_2_out_matrices_test",
+	"version": "1",
+	"author": "doc.ai",
+	"license": "",
+	"model": {
+		"file": "model.tflite",
+		"quantized": false
+	},
+	"inputs": [
+		{
+			"name": "input1",
+			"type": "array",
+			"shape": [4,4]
+		},
+		{
+			"name": "input2",
+			"type": "array",
+			"shape": [4,4]
+		}
+	],
+	"outputs": [
+		{
+			"name": "output1",
+			"type": "array",
+			"shape": [4,4]
+		},
+		{
+			"name": "output2",
+			"type": "array",
+			"shape": [4,4]
+		}
+	]
+}
