+++ tflite/src/androidTest/assets/mobilenet_v1_1.0_224_quant.tiobundle/model.json
@@ -1,0 +1,36 @@
+{
+	"name": "MobileNet V1 1.0 224 Quantized",
+	"details": "MobileNet V1 with a width multiplier of 1.0 and an input resolution of 224x224. Quantized.\n\nMobileNets are based on a streamlined architecture that have depth-wise separable convolutions to build light weight deep neural networks. Trained on ImageNet with categories such as trees, animals, food, vehicles, person etc. MobileNets: Efficient Convolutional Neural Networks for Mobile Vision Applications.",
+	"id": "mobilenet-v1-100-224-quantized",
+	"version": "1",
+	"author": "Andrew G. Howard, Menglong Zhu, Bo Chen, Dmitry Kalenichenko, Weijun Wang, Tobias Weyand, Marco Andreetto, Hartwig Adam",
+	"license": "Apache License. Version 2.0 http://www.apache.org/licenses/LICENSE-2.0",
+	"model": {
+		"file": "mobilenet_v1_1.0_224_quant.tflite",
+		"quantized": true,
+		"backend": "tflite",
+		"modes": ["predict"]
+	},
+	"inputs": [
+		{
+			"name": "image",
+			"type": "image",
+			"shape": [224,224,3],
+			"format": "RGB"
+		}
+	],
+	"outputs": [
+		{
+			"name": "classification",
+			"type": "array",
+			"shape": [1,1001],
+			"labels": "labels.txt",
+			"dequantize": {
+				"standard": "[0,1]"
+			}
+		}
+	],
+	"options": {
+		"device_position": "back"
+	}
+}
